refactor(auth): Relocate unauthorized page
